+++ engine/app/components/citizens_advice_cookie_preferences/success_message.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module CitizensAdviceCookiePreferences
+  class SuccessMessage < ViewComponent::Base
+    attr_reader :message, :return_url
+
+    def initialize(message:, return_url: nil)
+      super()
+      @message = message
+      @return_url = return_url
+    end
+
+    def render?
+      message.present?
+    end
+  end
+end